LPS-201728 Make the ReindexPortalBackgroundTaskExecutor an OSGi component as well
