+++ modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/background/task/ReindexPortalBackgroundTaskExecutor.java
@@ -10,6 +10,7 @@ import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.backgroundtask.BackgroundTaskExecutor;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.module.service.Snapshot;
 import com.liferay.portal.kernel.search.background.task.ReindexBackgroundTaskConstants;
 import com.liferay.portal.kernel.search.background.task.ReindexStatusMessageSenderUtil;
 import com.liferay.portal.search.index.ConcurrentReindexManager;
@@ -17,30 +18,28 @@ import com.liferay.portal.search.index.SyncReindexManager;
 import com.liferay.portal.search.internal.SearchEngineInitializer;
 
 import org.osgi.framework.BundleContext;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Andrew Betts
  */
+@Component(
+	property = "background.task.executor.class.name=com.liferay.portal.search.internal.background.task.ReindexPortalBackgroundTaskExecutor",
+	service = BackgroundTaskExecutor.class
+)
 public class ReindexPortalBackgroundTaskExecutor
 	extends BaseReindexBackgroundTaskExecutor {
 
-	public ReindexPortalBackgroundTaskExecutor(
-		BundleContext bundleContext,
-		ConcurrentReindexManager concurrentReindexManager,
-		PortalExecutorManager portalExecutorManager,
-		SyncReindexManager syncReindexManager) {
-
-		_bundleContext = bundleContext;
-		_concurrentReindexManager = concurrentReindexManager;
-		_portalExecutorManager = portalExecutorManager;
-		_syncReindexManager = syncReindexManager;
-	}
-
 	@Override
 	public BackgroundTaskExecutor clone() {
-		return new ReindexPortalBackgroundTaskExecutor(
-			_bundleContext, _concurrentReindexManager, _portalExecutorManager,
-			_syncReindexManager);
+		return this;
+	}
+
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_bundleContext = bundleContext;
 	}
 
 	@Override
@@ -63,9 +62,10 @@ public class ReindexPortalBackgroundTaskExecutor
 			try {
 				SearchEngineInitializer searchEngineInitializer =
 					new SearchEngineInitializer(
-						_bundleContext, companyId, _concurrentReindexManager,
-						executionMode, _portalExecutorManager,
-						_syncReindexManager);
+						_bundleContext, companyId,
+						_concurrentReindexManagerSnapshot.get(), executionMode,
+						_portalExecutorManager,
+						_syncReindexManagerSnapshot.get());
 
 				searchEngineInitializer.reindex();
 			}
@@ -90,9 +90,18 @@ public class ReindexPortalBackgroundTaskExecutor
 	private static final Log _log = LogFactoryUtil.getLog(
 		ReindexPortalBackgroundTaskExecutor.class);
 
-	private final BundleContext _bundleContext;
-	private final ConcurrentReindexManager _concurrentReindexManager;
-	private final PortalExecutorManager _portalExecutorManager;
-	private final SyncReindexManager _syncReindexManager;
+	private static final Snapshot<ConcurrentReindexManager>
+		_concurrentReindexManagerSnapshot = new Snapshot<>(
+			ReindexPortalBackgroundTaskExecutor.class,
+			ConcurrentReindexManager.class, null, true);
+	private static final Snapshot<SyncReindexManager>
+		_syncReindexManagerSnapshot = new Snapshot<>(
+			ReindexPortalBackgroundTaskExecutor.class, SyncReindexManager.class,
+			null, true);
+
+	private BundleContext _bundleContext;
+
+	@Reference
+	private PortalExecutorManager _portalExecutorManager;
 
 }
